requirements.txt
aiofiles
bs4
certifi
dnspython
pyrogram
httpx
markdown
motor
gtts
psutil
hachoir
googletrans==4.0.0-rc1
cachetools
tinytag
pyromod
python-dotenv
pyyaml
aiohttp
requests
telegraph
tgcrypto
ujson
natsort
gitpython
pillow
ffmpeg-python
